migration to c++ ready to be tested!

+++ Inc/pipicofx/delayMemoryHandler.h
@@ -8,7 +8,7 @@ typedef struct
 } delayMemoryBlockType;
 
 void initDelayMemoryHandler(void);
-int16_t * mallocDelayMemory(uint16_t size);
+int16_t * mallocDelayMemory(uint32_t size);
 void freeDelayMemory(int16_t * ptr);
 void setStart(uint32_t strt); // only used for testing on a 64bit system
 #endif
